requirements.txt
bottle==0.13.2
bottle-websocket==0.2.9
Eel==0.17.0
future==1.0.0
gevent==24.10.3
gevent-websocket==0.10.1
greenlet==3.1.1
numpy==2.1.3
pandas==2.2.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.3.0
six==1.16.0
tzdata==2024.2
whichcraft==0.6.1
zope.event==5.0
zope.interface==7.1.1
google-genai